Integrated MPI test code.
